pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "utils"
version = "0.1.0"
description = "Shared utilities for Django, FastAPI, Alpaca, AWS, data, and ML."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "Cole Dylewski" }
]
keywords = ["utilities", "fastapi", "aws", "alpaca", "data", "ml", "snowflake"]
# All dependencies are required - no optional dependencies
dependencies = [
  # Alpaca Trading API - NOTE: Using custom REST API implementation, SDK not needed
  # "alpaca-trade-api>=3.1.0",  # Removed - using REST APIs instead of SDK
  # AWS Services
  "boto3>=1.28",
  # Note: botocore is included with boto3, no need to specify separately
  # Database
  "SQLAlchemy>=2.0",
  "psycopg[binary]>=3.1",
  "psycopg2-binary>=2.9.0",  # Required for legacy code using psycopg2
  "asyncpg",
  # FastAPI
  "fastapi>=0.110",
  "pydantic>=2.0",
  "uvicorn>=0.23",
  # HTTP Clients
  "httpx>=0.24.0",
  "requests>=2.32.0",
  # urllib3>=2.0 required by tableauserverclient>=0.38
  "urllib3>=2.0.0,<3.0",
  # Machine Learning
  "numpy>=1.26",
  "pandas>=2.0",
  "scikit-learn>=1.4",
  # Tableau
  # Use version compatible with urllib3>=2.0 (tableauserverclient 0.38+ requires urllib3==2.x)
  "tableauserverclient>=0.38",
  # Testing
  "pytest>=7.4.0",
  "pytest-cov>=4.1.0",
  "pytest-asyncio>=0.21.0",
  "pytest-mock>=3.12.0",
  "pytest-timeout>=2.2.0",
  "pytest-benchmark>=4.0.0",
  "pytest-xdist>=3.3.0",
  "moto>=5.0.0",
  "responses>=0.24.0",
  "freezegun>=1.2.0",
  "faker>=20.0.0",
  # Code Quality
  "ruff>=0.1.0",
  "mypy>=1.7.0",
  "pre-commit>=3.5.0",
  # Security Scanning
  "safety>=2.3.0",
  "bandit>=1.7.0",
  # Coverage Reporting
  "coverage>=7.3.0",
  # Build and Distribution
  "build>=1.0.0",
  "twine>=4.0.0",
  # Documentation
  "mkdocs>=1.5.0",
  "mkdocs-material>=9.0.0",
  "mkdocstrings[python]>=0.23.0",
  # Vault Integration
  "hvac>=1.0.0",
  # CLI and Logging
  "click>=8.1.0",
  "structlog>=23.0.0",
  # Additional Utilities
  "redis>=4.0.0",  # Redis client for caching and utilities
  "nest_asyncio>=1.5.0",  # For running async code in sync contexts
]

# Optional dependencies kept for backward compatibility but all are now in base dependencies
[project.optional-dependencies]

[project.urls]
Homepage = "https://github.com/Cole-Dylewski/_utils"

[project.scripts]
utils = "cli:main"

[tool.setuptools.packages.find]
where = ["python"]
include = ["*"]
exclude = ["tests*"]
namespaces = false

[tool.pytest.ini_options]
testpaths = ["python/tests"]
pythonpath = ["python"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
  "-v",
  "--strict-markers",
  "--strict-config",
  "--tb=short",
    "--cov=python",
  "--cov-report=term-missing",
  "--cov-report=html",
  "--cov-report=xml",
  "--cov-branch",
    "--cov-fail-under=80",
  "--durations=10",
  "-ra",
]
markers = [
  "unit: Unit tests",
  "integration: Integration tests",
  "slow: Slow running tests",
  "aws: Tests requiring AWS services",
  "alpaca: Tests requiring Alpaca API",
  "vault: Tests requiring Vault",
  "terraform: Tests requiring Terraform",
  "ansible: Tests requiring Ansible",
  "requires_network: Tests requiring network access",
  "requires_docker: Tests requiring Docker",
]

[tool.coverage.run]
source = ["python"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/__init__.py",
    "*/setup.py",
    "run_ci.py",
    "setup-venv.py",
    "tasks/*",
]

[tool.coverage.report]
exclude_lines = [
  "pragma: no cover",
  "def __repr__",
  "raise AssertionError",
  "raise NotImplementedError",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
  "@abstractmethod",
  "@abc.abstractmethod",
]
precision = 2
show_missing = true
skip_covered = false

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
  "E", "W", "F", "I", "N", "UP", "B", "C4", "DTZ", "T10", "EM", "ISC",
  "ICN", "PIE", "PT", "Q", "RSE", "RET", "SIM", "TID", "ARG", "PTH",
  "ERA", "PD", "PGH", "PL", "TRY", "NPY", "RUF",
]
ignore = [
  "E501",
  "PLR0913",
  "PLR2004",
  "TRY003",
  "EM101",
  "EM102",
  "ISC001",
]
fixable = ["ALL"]
unfixable = []

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"python/tests/**/*.py" = ["ARG", "S101"]

[tool.ruff.isort]
known-first-party = ["_utils"]
force-sort-within-sections = true
split-on-trailing-comma = true

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true
show_error_codes = true
ignore_missing_imports = true
